Empty layers cast output as floats to prevent conflicts

+++ layers/empty_replacement.py
@@ -5,4 +5,4 @@ class EmptyReplacement(tf.keras.layers.Layer):
         super(EmptyReplacement, self).__init__()
 
     def call(self, inputs):
-        return tf.expand_dims(inputs, -1)
+        return tf.cast(tf.expand_dims(inputs, -1), tf.float32)
